\name{iontreeData-package}
\alias{iontreeData-package}
\alias{iontreeData}
\docType{package}
\title{
Data for iontree package  
}
\description{
There are 3 data sets including raw MS2, MS3 scans 'MS2RAW', 'MS3RAW' and a SQLite-based ion tree database named 'mzDB'.
}
\details{
\tabular{ll}{
Package: \tab iontreeData\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
License: \tab GPL-2 \cr
LazyLoad: \tab yes\cr
 }
}

\author{
Mingshu Cao <mingshu.cao@agresearch.co.nz>
}
\references{
Cao M, Koulman A, Johnson LJ, Lane GA and Rasmussen S. 2008. Plant Physiology 146:4
}

\keyword{ package }
\seealso{
}
\examples{
}
