\name{MS2RAW}
\alias{MS2RAW}
\docType{data}
\title{
All the MS2 scans from 4 samples
}
\description{
MS2RAW data from direct infusion mass spectrometry.
}
\usage{data(MS2RAW)}
\format{
  see examples
}
\details{
}
\source{
Cao M, Koulman A, Johnson LJ, Lane GA and Rasmussen S. 2008. Plant Physiology 146:4
}
\references{
}
\examples{
data(MS2RAW)
length(MS2RAW)
attributes(MS2RAW[[1]])
}
\keyword{datasets}
