\name{MS3RAW}
\alias{MS3RAW}
\docType{data}
\title{
All the MS3 scans from 4 samples
}
\description{
MS3RAW data from direct infusion mass spectrometry.
}
\usage{data(MS3RAW)}
\format{
  see examples
}
\details{
}
\source{
Cao M, Koulman A, Johnson LJ, Lane GA and Rasmussen S. 2008. Plant Physiology 146:4
}
\references{
}
\examples{
data(MS3RAW)
}
\keyword{datasets}
